2.2: Fix target prediction broken by Rhamphoryncus's patch.
git-svn-id: https://warzone2100.svn.sourceforge.net/svnroot/warzone2100/branches/2.2@7279 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/combat.c
@@ -347,11 +347,11 @@ void combFire(WEAPON *psWeap, BASE_OBJECT *psAttacker, BASE_OBJECT *psTarget, in
 
 			if (proj_Direct(psStats) || dist <= psStats->minRange)
 			{
-				flightTime = dist / psStats->flightSpeed;
+				flightTime = (double)dist / psStats->flightSpeed;
 			}
 			else
 			{
-				flightTime = sqrt(dist) / 30;  /* Purely a guess, but surprisingly effective */
+				flightTime = sqrt((double)dist) / 30;  /* Purely a guess, but surprisingly effective */
 			}
 
 			if (psTarget->lastHitWeapon == WSC_EMP)
@@ -364,7 +364,7 @@ void combFire(WEAPON *psWeap, BASE_OBJECT *psAttacker, BASE_OBJECT *psTarget, in
 				}
 				else
 				{
-					flightTime -= empTime / 1000;
+					flightTime -= (double)empTime / 1000;
 					if (flightTime < 0.0)
 					{
 						flightTime = 0.0;
